requirements.txt
flet>=0.80.0
pillow>=10.0.0
